UPSTREAM: stddef.h: Introduce sizeof_field()
The size of fields within a structure is needed in a few places in the
kernel already, and will be needed for the usercopy whitelisting when
declaring whitelist regions within structures. This creates a dedicated
macro and redefines offsetofend() to use it.

Existing usage, ignoring the 1200+ lustre assert uses:

$ git grep -E 'sizeof\(\(\((struct )?[a-zA-Z_]+ \*\)0\)->' | \
	grep -v staging/lustre | wc -l
65

Change-Id: I1ebbaa5aa32f59b2e5a0b376f11d120e71815941
Signed-off-by: Kees Cook <keescook@chromium.org>
(cherry picked from commit 4229a470175be14e1d2648713be8a5e8e8fbea02)